khedam5

+++ src/main/java/ma/xproce/backend/service/EvaluationService.java
@@ -3,9 +3,12 @@ package ma.xproce.backend.service;
 import ma.xproce.backend.Dao.entities.Evaluation;
 import ma.xproce.backend.Dao.repositories.EvaluationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class EvaluationService {
@@ -13,28 +16,32 @@ public class EvaluationService {
     @Autowired
     private EvaluationRepository evaluationRepository;
 
-    // Créer une nouvelle évaluation
     public Evaluation createEvaluation(Evaluation evaluation) {
         return evaluationRepository.save(evaluation);
     }
 
-    // Obtenir toutes les évaluations d'un article
-    public List<Evaluation> getEvaluationsByArticle(Long articleId) {
-        return evaluationRepository.findByArticleId(articleId);
+    public Page<Evaluation> searchEvaluations(String keyword, int page, int size) {
+        return evaluationRepository.findByCommentsContainingIgnoreCase(keyword, PageRequest.of(page, size));
     }
 
-    // Obtenir toutes les évaluations faites par un reviewer
-    public List<Evaluation> getEvaluationsByReviewer(Long reviewerId) {
-        return evaluationRepository.findByReviewerId(reviewerId);
+    public Evaluation getEvaluationById(Long id) {
+        Optional<Evaluation> optionalEvaluation = evaluationRepository.findById(id);
+        return optionalEvaluation.orElse(null);
     }
 
-    // Mettre à jour une évaluation
     public Evaluation updateEvaluation(Evaluation evaluation) {
         return evaluationRepository.save(evaluation);
     }
 
-    // Supprimer une évaluation
-    public void deleteEvaluation(Long id) {
-        evaluationRepository.deleteById(id);
+    public boolean deleteEvaluationById(Long id) {
+        if (evaluationRepository.existsById(id)) {
+            evaluationRepository.deleteById(id);
+            return true;
+        }
+        return false;
+    }
+
+    public List<Evaluation> getAllEvaluations() {
+        return evaluationRepository.findAll();
     }
 }
